Pull out SemVer to own file
